test: refactor tests
Fixes: https://github.com/metarhia/jstp/issues/181
