+++ test/node/message-parser.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const tap = require('tap');
+const test = require('tap').test;
 const jstp = require('../..');
 
 const testCases = require('../fixtures/message-parser');
@@ -8,8 +8,9 @@ const testCases = require('../fixtures/message-parser');
 testCases.forEach((testCase) => {
   const result = [];
   const remainder = jstp.parseNetworkPackets(testCase.message, result);
-  tap.strictSame(result, testCase.result,
-    `must properly parse ${testCase.name}`);
-  tap.strictSame(remainder, testCase.remainder,
-    `must leave a valid remainder after parsing ${testCase.name}`);
+  test(`must properly parse ${testCase.name}`, (test) => {
+    test.strictSame(result, testCase.result);
+    test.strictSame(remainder, testCase.remainder);
+    test.end();
+  });
 });
