+++ test/node/remote-error-default-messages.js
@@ -1,13 +1,15 @@
 'use strict';
 
-const tap = require('tap');
+const test = require('tap').test;
 
-const jstp = require('../../');
+const jstp = require('../..');
 const RemoteError = jstp.RemoteError;
 
 const expectedMessages = Object.keys(jstp)
   .filter(key => key.startsWith('ERR_'))
   .map(key => jstp[key].toString());
 
-tap.includes(Object.keys(RemoteError.defaultMessages), expectedMessages,
-  'Must have a default message for every predefined error');
+test('Must have a default message for every predefined error', (test) => {
+  test.includes(Object.keys(RemoteError.defaultMessages), expectedMessages);
+  test.end();
+});
